refactor(server): extract CLI commands into separate modules
Reduces cli.js from 1084 lines to 41 lines by extracting each command
group into its own module under src/cli/. Shared options (--config,
--model, --tunnel, etc.) are defined once in cli/shared.js and reused
by both start and dev commands via addServerOptions().

Refs #1842
